refactor hles transformation tests
