+++ dap-etl/transformation/src/test/scala/org/broadinstitute/monster/dap/dog/DogTransformationsSpec.scala
@@ -1,12 +1,11 @@
-package org.broadinstitute.monster.dap
+package org.broadinstitute.monster.dap.dog
 
 import org.broadinstitute.monster.dap.common.RawRecord
-import org.broadinstitute.monster.dap.dog.DogTransformations
 import org.broadinstitute.monster.dogaging.jadeschema.fragment._
 import org.broadinstitute.monster.dogaging.jadeschema.table.HlesDog
+import org.scalatest.OptionValues
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
-import org.scalatest.OptionValues
 
 class DogTransformationsSpec extends AnyFlatSpec with Matchers with OptionValues {
   behavior of "DogTransformations"
